+++ lib/rational.dart
@@ -14,38 +14,30 @@
 
 library rational;
 
-import 'package:rational/bigint.dart';
-
-final IS_JS = identical(1, 1.0);
-
 final _PATTERN = new RegExp(r"^([+-]?\d+)(\.\d+)?([eE][+-]?\d+)?$");
 
-final _0 = new Rational<dynamic>(0);
-final _1 = new Rational<dynamic>(1);
-final _5 = new Rational<dynamic>(5);
-final _10 = new Rational<dynamic>(10);
+final _0 = new Rational.fromInt(0);
+final _1 = new Rational.fromInt(1);
+final _5 = new Rational.fromInt(5);
+final _10 = new Rational.fromInt(10);
 
-dynamic _int(int value) => IS_JS ? new BigInt.fromJsInt(value) : value;
-final _INT_0 = _int(0);
-final _INT_1 = _int(1);
-final _INT_2 = _int(2);
-final _INT_5 = _int(5);
-final _INT_10 = _int(10);
-final _INT_31 = _int(31);
+final _INT_0 = new BigInt.from(0);
+final _INT_1 = new BigInt.from(1);
+final _INT_2 = new BigInt.from(2);
+final _INT_5 = new BigInt.from(5);
+final _INT_10 = new BigInt.from(10);
+final _INT_31 = new BigInt.from(31);
 
-dynamic _parseInt(String text) => IS_JS ? BigInt.parse(text) : int.parse(text);
-
-dynamic _gcd(dynamic a, dynamic b) {
+BigInt _gcd(BigInt a, BigInt b) {
   while (b != _INT_0) {
-    var t = b;
+    final t = b;
     b = a % t;
     a = t;
   }
   return a;
 }
 
-abstract class Rational<T extends dynamic /*int|BigInt*/ >
-    implements Comparable<Rational> {
+class Rational implements Comparable<Rational> {
   static Rational parse(String decimalValue) {
     final match = _PATTERN.firstMatch(decimalValue);
     if (match == null) {
@@ -61,12 +53,12 @@ abstract class Rational<T extends dynamic /*int|BigInt*/ >
       for (int i = 1; i < group2.length; i++) {
         denominator = denominator * _INT_10;
       }
-      numerator = _parseInt('${group1}${group2.substring(1)}');
+      numerator = BigInt.parse('${group1}${group2.substring(1)}');
     } else {
-      numerator = _parseInt(group1);
+      numerator = BigInt.parse(group1);
     }
     if (group3 != null) {
-      var exponent = _parseInt(group3.substring(1));
+      var exponent = BigInt.parse(group3.substring(1));
       while (exponent > _INT_0) {
         numerator = numerator * _INT_10;
         exponent -= _INT_1;
@@ -76,22 +68,15 @@ abstract class Rational<T extends dynamic /*int|BigInt*/ >
         exponent += _INT_1;
       }
     }
-    return new Rational._normalize(numerator, denominator);
+    return new Rational(numerator, denominator);
   }
 
-  final T _numerator, _denominator;
+  factory Rational.fromInt(int numerator, [int denominator = 1]) =>
+      new Rational(new BigInt.from(numerator), new BigInt.from(denominator));
 
-  Rational._(this._numerator, this._denominator);
-
-  factory Rational._normalized(numerator, denominator) => IS_JS
-      ? new _RationalJs._normalized(numerator, denominator)
-      : new _RationalVM._normalized(numerator, denominator);
-
-  factory Rational(int numerator, [int denominator = 1]) =>
-      new Rational._normalize(_int(numerator), _int(denominator));
-
-  factory Rational._normalize(numerator, denominator) {
-    if (denominator == _INT_0) throw new IntegerDivisionByZeroException();
+  factory Rational(BigInt numerator, [BigInt denominator]) {
+    denominator ??= _INT_1;
+    if (denominator == _INT_0) throw new ArgumentError();
     if (numerator == _INT_0) return new Rational._normalized(_INT_0, _INT_1);
     if (denominator < _INT_0) {
       numerator = -numerator;
@@ -104,28 +89,29 @@ abstract class Rational<T extends dynamic /*int|BigInt*/ >
         ? new Rational._normalized(numerator, denominator)
         : new Rational._normalized(numerator ~/ gcd, denominator ~/ gcd);
   }
+  Rational._normalized(this.numerator, this.denominator);
+  // : assert(numerator != null),
+  //   assert(denominator != null),
+  //   assert(denominator > _INT_0),
+  //   assert(_gcd(numerator.abs(), denominator) == _INT_1);
 
-  @Deprecated('can give bad value with dart2js')
-  int get numerator;
+  final BigInt numerator, denominator;
 
-  @Deprecated('can give bad value with dart2js')
-  int get denominator;
+  bool get isInteger => denominator == _INT_1;
 
-  bool get isInteger => _denominator == _INT_1;
-
-  int get hashCode => (_numerator + _INT_31 * _denominator).hashCode;
+  int get hashCode => (numerator + _INT_31 * denominator).hashCode;
 
   bool operator ==(Object other) =>
       other is Rational &&
-      _numerator == other._numerator &&
-      _denominator == other._denominator;
+      numerator == other.numerator &&
+      denominator == other.denominator;
 
   String toString() {
-    if (_numerator == _INT_0) return '0';
+    if (numerator == _INT_0) return '0';
     if (isInteger)
-      return '$_numerator';
+      return '$numerator';
     else
-      return '$_numerator/$_denominator';
+      return '$numerator/$denominator';
   }
 
   String toDecimalString() {
@@ -141,24 +127,24 @@ abstract class Rational<T extends dynamic /*int|BigInt*/ >
   }
   // implementation of Comparable
 
-  int compareTo(Rational other) => (_numerator * other._denominator)
-      .compareTo(other._numerator * _denominator);
+  int compareTo(Rational other) =>
+      (numerator * other.denominator).compareTo(other.numerator * denominator);
 
   // implementation of num
 
   /** Addition operator. */
-  Rational operator +(Rational other) => new Rational._normalize(
-      _numerator * other._denominator + other._numerator * _denominator,
-      _denominator * other._denominator);
+  Rational operator +(Rational other) => new Rational(
+      numerator * other.denominator + other.numerator * denominator,
+      denominator * other.denominator);
 
   /** Subtraction operator. */
-  Rational operator -(Rational other) => new Rational._normalize(
-      _numerator * other._denominator - other._numerator * _denominator,
-      _denominator * other._denominator);
+  Rational operator -(Rational other) => new Rational(
+      numerator * other.denominator - other.numerator * denominator,
+      denominator * other.denominator);
 
   /** Multiplication operator. */
-  Rational operator *(Rational other) => new Rational._normalize(
-      _numerator * other._numerator, _denominator * other._denominator);
+  Rational operator *(Rational other) => new Rational(
+      numerator * other.numerator, denominator * other.denominator);
 
   /** Euclidean modulo operator. */
   Rational operator %(Rational other) {
@@ -168,8 +154,8 @@ abstract class Rational<T extends dynamic /*int|BigInt*/ >
   }
 
   /** Division operator. */
-  Rational operator /(Rational other) => new Rational._normalize(
-      _numerator * other._denominator, _denominator * other._numerator);
+  Rational operator /(Rational other) => new Rational(
+      numerator * other.denominator, denominator * other.numerator);
 
   /**
    * Truncating division operator.
@@ -180,7 +166,7 @@ abstract class Rational<T extends dynamic /*int|BigInt*/ >
   Rational operator ~/(Rational other) => (this / other).truncate();
 
   /** Negate operator. */
-  Rational operator -() => new Rational._normalized(-_numerator, _denominator);
+  Rational operator -() => new Rational._normalized(-numerator, denominator);
 
   /** Return the remainder from dividing this [num] by [other]. */
   Rational remainder(Rational other) => this - (this ~/ other) * other;
@@ -199,7 +185,7 @@ abstract class Rational<T extends dynamic /*int|BigInt*/ >
 
   bool get isNaN => false;
 
-  bool get isNegative => _numerator < _INT_0;
+  bool get isNegative => numerator < _INT_0;
 
   bool get isInfinite => false;
 
@@ -241,7 +227,8 @@ abstract class Rational<T extends dynamic /*int|BigInt*/ >
    * Returns the integer value obtained by discarding any fractional
    * digits from this [num].
    */
-  Rational truncate() => new Rational._normalized(_toInt(), _INT_1);
+  Rational truncate() =>
+      new Rational._normalized(numerator ~/ denominator, _INT_1);
 
   /**
    * Returns the integer value closest to `this`.
@@ -284,12 +271,13 @@ abstract class Rational<T extends dynamic /*int|BigInt*/ >
 
   /**
    * Truncates this [num] to an integer and returns the result as an [int].
-   *
-   * **WARNING for dart2js** : It can give bad result for large number.
    */
-  int toInt();
+  int toInt() => toBigInt().toInt();
 
-  T _toInt() => _numerator ~/ _denominator as T;
+  /**
+   * Truncates this [num] to a big integer and returns the result as an [BigInt].
+   */
+  BigInt toBigInt() => numerator ~/ denominator;
 
   /**
    * Return this [num] as a [double].
@@ -298,16 +286,16 @@ abstract class Rational<T extends dynamic /*int|BigInt*/ >
    * approximation is returned. For numerically large integers, the
    * approximation may be infinite.
    */
-  double toDouble();
+  double toDouble() => numerator / denominator;
 
   /**
    * Inspect if this [num] has a finite precision.
    */
   bool get hasFinitePrecision {
     // the denominator should only be a product of powers of 2 and 5
-    var den = _denominator;
-    while (den % _INT_5 == _INT_0) den = den ~/ _INT_5 as T;
-    while (den % _INT_2 == _INT_0) den = den ~/ _INT_2 as T;
+    var den = denominator;
+    while (den % _INT_5 == _INT_0) den = den ~/ _INT_5;
+    while (den % _INT_2 == _INT_0) den = den ~/ _INT_2;
     return den == _INT_1;
   }
 
@@ -324,9 +312,9 @@ abstract class Rational<T extends dynamic /*int|BigInt*/ >
     if (!hasFinitePrecision) {
       throw new StateError("This number has an infinite precision: $this");
     }
-    var x = _numerator;
-    while (x % _denominator != _INT_0) x *= _INT_10;
-    x = x ~/ _denominator as T;
+    var x = numerator;
+    while (x % denominator != _INT_0) x *= _INT_10;
+    x = x ~/ denominator;
     return x.abs().toString().length;
   }
 
@@ -344,8 +332,8 @@ abstract class Rational<T extends dynamic /*int|BigInt*/ >
       throw new StateError("This number has an infinite precision: $this");
     }
     var i = 0;
-    var x = _numerator;
-    while (x % _denominator != _INT_0) {
+    var x = numerator;
+    while (x % denominator != _INT_0) {
       i++;
       x *= _INT_10;
     }
@@ -358,19 +346,19 @@ abstract class Rational<T extends dynamic /*int|BigInt*/ >
    */
   String toStringAsFixed(int fractionDigits) {
     if (fractionDigits == 0) {
-      return round()._toInt().toString();
+      return round().toBigInt().toString();
     } else {
       var mul = _INT_1;
       for (int i = 0; i < fractionDigits; i++) mul *= _INT_10;
-      final mulRat = new Rational._normalize(mul, _INT_1);
+      final mulRat = new Rational(mul);
       final lessThanOne = abs() < _1;
       final tmp = (lessThanOne ? (abs() + _1) : abs()) * mulRat;
       final tmpRound = tmp.round();
       final intPart =
           (lessThanOne ? ((tmpRound ~/ mulRat) - _1) : (tmpRound ~/ mulRat))
-              ._toInt();
+              .toBigInt();
       final decimalPart =
-          tmpRound._toInt().toString().substring(intPart.toString().length);
+          tmpRound.toBigInt().toString().substring(intPart.toString().length);
       return '${isNegative ? '-' : ''}${intPart}.${decimalPart}';
     }
   }
@@ -388,25 +376,4 @@ abstract class Rational<T extends dynamic /*int|BigInt*/ >
    */
   String toStringAsPrecision(int precision) =>
       toDouble().toStringAsPrecision(precision);
-}
-
-class _RationalJs extends Rational<BigInt> {
-  _RationalJs._normalized(BigInt numerator, BigInt denominator)
-      : super._(numerator, denominator);
-
-  int get numerator => int.parse('$_numerator');
-  int get denominator => int.parse('$_denominator');
-  int toInt() => int.parse(_toInt().toString());
-  double toDouble() =>
-      double.parse('$_numerator') / double.parse('$_denominator');
-}
-
-class _RationalVM extends Rational<int> {
-  _RationalVM._normalized(int numerator, int denominator)
-      : super._(numerator, denominator);
-
-  int get numerator => _numerator;
-  int get denominator => _denominator;
-  int toInt() => _numerator ~/ _denominator;
-  double toDouble() => _numerator / _denominator;
 }